- change error detection conditions in AdminController

+++ src/main/java/org/example/service/admin/AdminServiceImplementation.java
@@ -4,6 +4,7 @@ import org.example.entity.Admin;
 import org.example.entity.User;
 import org.example.model.Response;
 import org.example.repository.admin.AdminRepository;
+import org.example.repository.user.UserRepository;
 import org.example.service.ValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,12 +13,14 @@ import org.springframework.stereotype.Service;
 public class AdminServiceImplementation implements AdminService{
 
     private final AdminRepository repository;
+    private final UserRepository userRepository;
 
     private ValidationService validationService;
     @Autowired
-    public AdminServiceImplementation(AdminRepository repository) {
+    public AdminServiceImplementation(AdminRepository repository, UserRepository userRepository) {
         validationService=new ValidationService();
         this.repository = repository;
+        this.userRepository = userRepository;
     }
 
     /**
@@ -52,6 +55,10 @@ public class AdminServiceImplementation implements AdminService{
      */
     @Override
     public Response<Boolean> removeAdmin(int adminID, String adminEmail) {
+        Response isAdminDataCorrect = userRepository.getUser(adminID, adminEmail);
+
+        if(isAdminDataCorrect.getObjectToBeReturned()==null)
+            return new Response<>("admin data incorrect", 404, true, true, false);
         return repository.removeAdmin(adminID, adminEmail);
     }
 
@@ -60,6 +67,10 @@ public class AdminServiceImplementation implements AdminService{
      */
     @Override
     public Response<Boolean> deactivateCustomer(int customerID, String customerEmail) {
+        Response isAdminDataCorrect = userRepository.getUser(customerID, customerEmail);
+
+        if(isAdminDataCorrect.getObjectToBeReturned()==null)
+            return new Response<>("customer data incorrect", 404, true, true, false);
         return repository.deactivateCustomer(customerID, customerEmail);
     }
